+++ src/games/slashchain/components/hands.tsx
